Move to ns-based output times and frequencies
This is a big change. We have moved away from the tsc-based
timing and to a ns-based timing.

The user now specifies a sample frequency, not a TSC mask.

We no longer need to worry about -O0, since the count is now
volatile.

We automatically do internal computations of the time base.

It is a ton easier to work with now. It is providing results
compatible with the previous code, so I don't think I broke anything.

It needs some fixes to work on Akaros, but they are pretty minor.

Included is showfft.m, for octave. Usage:
showfft("file.dat", frequency);

Please test.

Signed-off-by: Ronald G. Minnich <rminnich@google.com>