Make the main loops / ftqcore OS-independent
The main part of the benchmark will be the same regardless of OS.  I noticed
slight variations in the main loop based on the version of GCC used.  By using
the same compiler, we can more easily compare different OSs on the same
hardware.

We're assuming that you have a native compiler that works for the architecture
you're running on (e.g. x86_64), and that the calling convention is the same
from Linux to Akaros.  Which it is for x86.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>