Make samples an array of structs
Instead of two longs.  The old style was harder to read and modify - say we
wanted to add another field to a sample.  The struct sample was implied in the
old code, and it is now explicit.

I was motivated to do this by seeing a strong 4 sample pattern: {710, 711, 710,
720}.  I suspected it was due to four samples fitting on a single cache line.
I wanted to check what happens if I increase the size of the sample to fit a
fresh CL (possibly implying a CL miss every sample, though the prefetcher might
defeat that).

Note that this change actually dropped performance by about 8% on my Haswell.
The culprit was a combination of a bunch of things.  Due to this change, the
compiler inserted 4 bytes for shr $0x4, %rax.  GCC wants to align loops at 16
byte boundaries (if it requires less than 10 bytes of padding), and otherwise
at 8 byte boundaries.  Due to those 4 bytes, the new version of the code for
main_loops had one extra nopl (8 bytes) in between the two inner most loops.
It also had slightly larger nopls in other places.

You'd think this might have caused more icache misses or something, but no -
the actual culprit was instruction decode stalls due to the instruction queue
being full (ILD_STALL:IQ_FULL).  10x more in the new code than the new, and
they were all in the main_loops.  The old code hardly had any.  This was
similar on both Linux and Akaros.

Moral of the story: slight changes in the benchmark can have drastic effects in
the raw performance, and those effects may be machine dependent.  Of course,
the raw performance isn't as important as consistency between runs on a given
machine.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>