Prevent excessive loop alignment
GCC normally aligns loop targets to 16, if it requires less than 10 bytes of
nops, and otherwise aligns to 8.  This introduces enough extra nopl bytes to
stall the instruction decoder.

For reference, before commit 32964268c991 ("Make samples an array of structs"),
we had about 35 M/s.  With that change, we had 350 M/s.  With this change, we
have about 190 M/s.  This change got rid of the extra nopl, but one remaining
nopl is a bit longer - enough to trigger some stalls it seems.

That number of stalls doesn't effect perf as much either.  The old version (35
M/s) got around 710.  The normal alignment (350 M/s) got around 650.  This
version gets around 707.

Arguably, raw performance doesn't matter as much as consistency, and this only
changes the raw performance (or so it seems).  And this effect may be only
noticeable on some machines, so we don't really need a machine-specific change.
Still, my hunch is this will help a lot of machines and may remove some
processor-specific 'interference.'

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>