Add an option to delay the start of the test
Each thread will spin for delay_msec.  This helps avoid some machine-dependent
noise associated with starting the test.

For instance, on one of my Haswells I'd get a strong 100 Hz signal for the
first 180 msec or so.  This was independent of the benchmark (present at 100 Hz
regardless of FTQ's sampling frequency) and independent of OS.  On quiet
machines, the signal was visible by manually looking at the data.  The only
thing I found so far to quiesce the signal was to halt at C1, but simply
decreasing the p-state to the max non-turbo ratio wasn't enough.  Other
Haswells didn't have this problem.

Anyway, with the -d option, you can avoid the initial pain.  It seems like that
was why we had the 1000 samples in the first place.  It just so happens that
1000 samples is only 100 msec at the default 10 KHz sampling rate, so it wasn't
enough for my machine.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>