Fix interval drift
The problem was that our data intervals would slowly drift.  For example, over
a hundred samples, our clock could be off by a usec.

Part of the problem is that our computed ticksperns is wrong - it's almost
always an underestimate.  You can set it manually with -T.

The other problem is that when the samples[] ticks are converted to nsec, they
used nspercycle, which is *also* computed and possibly different than
ticksperns.  If you manually provided ticksperns, it was almost certainly
different than nspercycle.  The samples were using the correct tick interval,
but we lost that on conversion.

The simplest thing is to remove nspercycle and only rely on either -T
ticksperns or a computed value.  In the case of recent versions of Akaros, this
value will actually come from the MSR.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>