Remove OS specific header files
The --include linux.h is a pain with some build systems.  We aren't
actually getting anything of value from those headers either - the
OS-specific interface is in ftq.h.  So let's just remove the OS headers.

Alternatively, we could do something like

 #if defined(OS_LINUX)
 #include "linux.h"
 #elif
 ...

and put that in os.h.  (But not in ftq.h, which is included by
ftqcore.c!)

Signed-off-by: Barret Rhoden <brho@google.com>